Several Fixed and Improvements

+++ app/controllers/admin_panel/settings_controller.rb
@@ -3,32 +3,39 @@ module AdminPanel
     layout 'admin'
 
     def show
+      @settings = Setting.first_or_initialize
       @payment_methods = PaymentMethod.all
       @enabled_payment_methods = EnabledPaymentMethod.pluck(:payment_method_id)
     end
 
     def update
-      # Clear existing enabled payment methods
-      EnabledPaymentMethod.delete_all
+      @settings = Setting.first_or_initialize
 
-      # Save the enabled payment methods
-      if params[:enabled_payment_methods].present?
-        params[:enabled_payment_methods].each do |payment_method_id|
-          EnabledPaymentMethod.create!(
-            payment_method_id: payment_method_id,
-            admin_id: current_admin.id # Ensure admin_id is set
-          )
+      if @settings.update(settings_params)
+        EnabledPaymentMethod.delete_all
+        params[:enabled_payment_methods]&.each do |payment_method_id|
+          EnabledPaymentMethod.create!(payment_method_id: payment_method_id, admin_id: current_admin.id)
         end
-      end
 
-      # Update payment method descriptions
-      if params[:payment_method_descriptions].present?
-        params[:payment_method_descriptions].each do |payment_method_id, description|
+        params[:payment_method_descriptions]&.each do |payment_method_id, description|
           PaymentMethod.find(payment_method_id).update(description: description)
         end
-      end
 
-      redirect_to admin_panel_settings_path, notice: 'Settings updated successfully.'
+        redirect_to admin_panel_settings_path, notice: 'Settings updated successfully.'
+      else
+        Rails.logger.error "Settings update failed: #{@settings.errors.full_messages.to_sentence}"
+        flash.now[:alert] = 'Failed to update settings. Please check the input fields.'
+        render :show
+      end
+    end
+
+    private
+
+    def settings_params
+      params.expect(
+        setting: [ :company_name, :registration_number, :phone, :company_email, :address, :city,
+        :company_postcode, :language, :currency, :notifications_email, :logo ]
+      )
     end
   end
 end
